% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubnt-discovery.R
\name{parse_ubnt_discovery_response}
\alias{parse_ubnt_discovery_response}
\title{Parser for Ubiquiti Discovery Protocol responses}
\usage{
parse_ubnt_discovery_response(a)
}
\arguments{
\item{a}{either a \code{raw} vector with the raw response byte data or a base64 encoded version of the same}
}
\value{
an \code{ubi_d} classed \code{list}  with device \code{name}, \code{model_long} and \code{model_short} model info, \code{firmware} string
station \code{esssid}, device \code{uptime} (if provided), device \code{macs} (MAC addresses) and
\code{ips} (IPv4 addresses) and any unprocessed fields as named raw vectors (named by their field code).
}
\description{
Takes a raw or base64 encoded Ubiquity Discovery Protocol response and turns it into
a list.
}
\examples{
paste0(
  c(
    "AQAAggIACiSkPH9XF8+QRUICAAokpDx+VxfAqAUBAQAGJKQ8flcXCgAEACylqg",
    "sADEFpclJvdXRlciBIUAwABkxBUC1IUA0AD1dDVGVsLVdpRmktNTcxNw4AAQMDA",
    "CJYTS5hcjcyNDAudjUuNS42LjE3NzYyLjEzMDUyOC4xNzU1EAAC5LI=",
  ), collaspe = ""
) -> base64_response

parse_ubnt_discovery_response(base64_response)

\dontrun{
x <- ubnt_discovery_probe("some ip address or host name")
if (!is.null(x)) parse_ubnt_discovery_response(x)
}
}
\seealso{
Other Ubiquity Discovery Protocol functions: \code{\link{ubnt_discovery_probe}}
}
\concept{Ubiquity Discovery Protocol functions}
