% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubnt-discovery.R
\name{print.ubnt_d}
\alias{print.ubnt_d}
\title{Printer for ubnt_d object}
\usage{
\method{print}{ubnt_d}(x, ...)
}
\arguments{
\item{x}{ubnt_d object}

\item{...}{ignored}
}
\description{
Printer for ubnt_d object
}
\keyword{internal}
