% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubnt-discovery.R
\name{ubnt_discovery_probe}
\alias{ubnt_discovery_probe}
\title{Send a Ubiquiti Discovery probe to a host with a discovery query payload and receive a response}
\usage{
ubnt_discovery_probe(host, port = 10001L, payload = c(1, 0, 0, 0))
}
\arguments{
\item{host}{host or IP address of target}

\item{port}{defaults to \code{10001L} but folks may be running it on others}

\item{payload}{the discovery protocol query payload. The default is known to work.}
}
\value{
raw vector
}
\description{
The \code{payload} is not fixed since it appears there may be other discovery
query payloads. Use \code{\link[=parse_ubnt_discovery_response]{parse_ubnt_discovery_response()}} to handle the the raw
result from this function.
}
\examples{
\dontrun{
x <- ubnt_discovery_probe("some ip address or host name")
if (!is.null(x)) parse_ubnt_discovery_response(x)
}
}
\seealso{
Other Ubiquity Discovery Protocol functions: \code{\link{parse_ubnt_discovery_response}}
}
\concept{Ubiquity Discovery Protocol functions}
