% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{udp_daytime}
\alias{udp_daytime}
\title{Send an request for the day/time}
\usage{
udp_daytime(host, port = 13L)
}
\arguments{
\item{host}{host or IP address of target}

\item{port}{defaults to \code{13L} (the standard \code{echo} port)}
}
\value{
character vector
}
\description{
\code{daytime} servers are also rly bad things to put
on the public internet but there are a few, still, of them as
of 2019. Find an IP and make a query!
}
\examples{
\dontrun{
udp_daytime("195.34.89.241")
}
}
\seealso{
Other insecure old internet service utility functions: \code{\link{udp_echo}},
  \code{\link{udp_qotd}}
}
\concept{insecure old internet service utility functions}
