% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{udp_qotd}
\alias{udp_qotd}
\title{Send an request for the QOTD}
\usage{
udp_qotd(host, port = 17L)
}
\arguments{
\item{host}{host or IP address of target}

\item{port}{defaults to \code{17L} (the standard \code{echo} port)}
}
\value{
character vector
}
\description{
Quote of the Day servers are also rly bad things to put
on the public internet but there are over 40K of them as
of 2019. Find an IP and make a query!
}
\examples{
\dontrun{
udp_qotd("119.48.167.199")
}
}
\seealso{
Other insecure old internet service utility functions: \code{\link{udp_daytime}},
  \code{\link{udp_echo}}
}
\concept{insecure old internet service utility functions}
