% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udp-payload.R
\name{udp_send_payload}
\alias{udp_send_payload}
\title{Send a UDP payload and return the response (if any)}
\usage{
udp_send_payload(host, port, payload)
}
\arguments{
\item{host}{host or ip address of target}

\item{port}{UDP port number}

\item{payload}{payload to send (raw bytes; \code{\link[=as.raw]{as.raw()}} will be called if not)}
}
\description{
Send a UDP payload and return the response (if any)
}
\note{
there's a hardcoded max response size of 4096 bytes for now #lazy
}
